+++ packages/block-editor/src/utils/index.js
@@ -1,4 +1,9 @@
 export { default as transformStyles } from './transform-styles';
 export * from './block-variation-transforms';
 export { default as getPxFromCssUnit } from './parse-css-unit-to-px';
-export { getValueFromVariable as __experimentalGetValueFromVariable } from './style-variable-resolution';
+export {
+	PRESET_METADATA,
+	findInPresetsBy,
+	getValueFromVariable,
+} from './style-variable-resolution';
+export { getTypographyFontSizeValue } from './typography';
