+++ packages/edit-site/src/components/global-styles/test/utils.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import { getPresetVariableFromValue, getValueFromVariable } from '../utils';
+import { getPresetVariableFromValue } from '../utils';
 
 describe( 'editor utils', () => {
 	const themeJson = {
@@ -102,104 +102,6 @@ describe( 'editor utils', () => {
 					);
 					expect( actual ).toBe( 'var:preset|color|primary' );
 				} );
-			} );
-		} );
-	} );
-
-	describe( 'getValueFromVariable', () => {
-		describe( 'when provided an invalid variable', () => {
-			it( 'returns the originally provided value', () => {
-				const actual = getValueFromVariable(
-					themeJson,
-					'root',
-					undefined
-				);
-
-				expect( actual ).toBe( undefined );
-			} );
-		} );
-
-		describe( 'when provided a preset variable', () => {
-			it( 'retrieves the correct preset value', () => {
-				const actual = getValueFromVariable(
-					themeJson,
-					'root',
-					'var:preset|color|primary'
-				);
-
-				expect( actual ).toBe( '#007cba' );
-			} );
-		} );
-
-		describe( 'when provided a custom variable', () => {
-			it( 'retrieves the correct custom value', () => {
-				const actual = getValueFromVariable(
-					themeJson,
-					'root',
-					'var(--wp--custom--color--secondary)'
-				);
-
-				expect( actual ).toBe( '#a65555' );
-			} );
-		} );
-
-		describe( 'when provided a dynamic reference', () => {
-			it( 'retrieves the referenced value', () => {
-				const stylesWithRefs = {
-					...themeJson,
-					styles: {
-						color: {
-							background: {
-								ref: 'styles.color.text',
-							},
-							text: 'purple-rain',
-						},
-					},
-				};
-				const actual = getValueFromVariable( stylesWithRefs, 'root', {
-					ref: 'styles.color.text',
-				} );
-
-				expect( actual ).toBe( stylesWithRefs.styles.color.text );
-			} );
-
-			it( 'returns the originally provided value where value is dynamic reference and reference does not exist', () => {
-				const stylesWithRefs = {
-					...themeJson,
-					styles: {
-						color: {
-							text: {
-								ref: 'styles.background.text',
-							},
-						},
-					},
-				};
-				const actual = getValueFromVariable( stylesWithRefs, 'root', {
-					ref: 'styles.color.text',
-				} );
-
-				expect( actual ).toBe( stylesWithRefs.styles.color.text );
-			} );
-
-			it( 'returns the originally provided value where value is dynamic reference', () => {
-				const stylesWithRefs = {
-					...themeJson,
-					styles: {
-						color: {
-							background: {
-								ref: 'styles.color.text',
-							},
-							text: {
-								ref: 'styles.background.text',
-							},
-						},
-					},
-				};
-				const actual = getValueFromVariable( stylesWithRefs, 'root', {
-					ref: 'styles.color.text',
-				} );
-
-				expect( actual ).toBe( stylesWithRefs.styles.color.text );
 			} );
 		} );
 	} );
